Add core Failure rich types
These are types meant to give rich semantic meaning to
our applications.

They are build first and foremost to offer convenient ways
of encoding what went wrong, both for the short term and
the long term.

Moreover, the types herein ought to be used to express even
more meaningful common functionality, like a systematic
translation to HTTP response codes, or the like.