Add scalafmt file to each subfolder to please Intellij
For unknown reason — for this project in particular — Intellij scalafmt plugin
does not pick up the config file in the root folder. the command-line-tool, and sbt
plugin do pick it up. Therefore, I had to do this to ease my life.

Will have to create a minimal project that reproduces this nonsense, and file a
bug report for the scalafmt Intellij plugin.