Use better-monadic-for plugin
Add test that should not compile if the plugin no longer works as expected (for whatever reason). It is infinitely better for a library to use this plugin since the bytecode it generates is more efficient, and safer, due to the runtime casts introduced by .withFilter calls. And all clients can keep using their sub-par scala for